+++ Sources/application/views/admin/list_dv_admin_view.php
@@ -28,7 +28,7 @@
                         <td align="center">
                             <a class="btn btn-default" href="<?php echo base_url();?>index.php/admin/edit_dv"><em class="fa fa-pencil"></em></a>
                             <br>
-                            <a class="btn btn-danger" href="<?php echo base_url();?>index.php/admin/delete_dv/<?php echo $row['id_dv'];?>"><em class="fa fa-trash"></em></a>
+                            <a class="btn btn-danger" onclick="return confirm('Bạn có muốn xóa không');" href="<?php echo base_url();?>index.php/admin/delete_dv/<?php echo $row['id_dv'];?>"><em class="fa fa-trash"></em></a>
                         </td>
                         <td><?php echo $row['id_dv'];?></td>
                         <td><?php echo $row['ten_dv'];?></td>
